requirements.txt
flask
Pillow
requests
cairosvg
psycopg2-binary